+++ soccer/src/soccer/robot_action_client/robot_action_client.cpp
@@ -1,26 +1,12 @@
-#include <functional>
-#include <future>
-#include <memory>
-#include <string>
-#include <sstream>
-
-#include <spdlog/spdlog.h>
-#include "rj_msgs/action/robot_move.hpp"
-
-#include <rj_utils/logging.hpp>
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_action/rclcpp_action.hpp"
+#include "robot_action_client.hpp"
 
 namespace robot_action_client
 {
-class RobotActionClient : public rclcpp::Node
-{
-public:
   using RobotMove = rj_msgs::action::RobotMove;
   using GoalHandleRobotMove = rclcpp_action::ClientGoalHandle<RobotMove>;
 
-  RobotActionClient()
-    : Node("planner",  rclcpp::NodeOptions{}
+  RobotActionClient::RobotActionClient()
+    : rclcpp::Node("robot_action_client_node",  rclcpp::NodeOptions{}
                                     .automatically_declare_parameters_from_overrides(true)
                                     .allow_undeclared_parameters(true))
   {
@@ -33,7 +19,7 @@ public:
       std::bind(&RobotActionClient::send_goal, this));
   }
 
-  void send_goal()
+  void RobotActionClient::send_goal()
   {
     using namespace std::placeholders;
 
@@ -60,11 +46,7 @@ public:
     this->client_ptr_->async_send_goal(goal_msg, send_goal_options);
   }
 
-private:
-  rclcpp_action::Client<RobotMove>::SharedPtr client_ptr_;
-  rclcpp::TimerBase::SharedPtr timer_;
-
-  void goal_response_callback(std::shared_future<GoalHandleRobotMove::SharedPtr> future)
+  void RobotActionClient::goal_response_callback(std::shared_future<GoalHandleRobotMove::SharedPtr> future)
   {
     auto goal_handle = future.get();
     if (!goal_handle) {
@@ -74,14 +56,14 @@ private:
     }
   }
 
-  void feedback_callback(
+  void RobotActionClient::feedback_callback(
     GoalHandleRobotMove::SharedPtr,
     const std::shared_ptr<const RobotMove::Feedback> feedback)
   {
       // TODO: fill in feedback from server-side
   }
 
-  void result_callback(const GoalHandleRobotMove::WrappedResult & result)
+  void RobotActionClient::result_callback(const GoalHandleRobotMove::WrappedResult & result)
   {
     switch (result.code) {
       case rclcpp_action::ResultCode::SUCCEEDED:
@@ -98,15 +80,5 @@ private:
     }
     SPDLOG_INFO("Result received: {}", result.result->is_done);
   }
-};  // class RobotActionClient
 
 }  // namespace robot_action_client
-
-int main(int argc, char** argv) {
-    rclcpp::init(argc, argv);
-    rj_utils::set_spdlog_default_ros2("processor");
-    SPDLOG_ERROR("Robot action client is being spun");
-    auto robot_action_client_node = std::make_shared<robot_action_client::RobotActionClient>();
-    /* start_global_param_provider(planner.get(), kGlobalParamServerNode); */
-    rclcpp::spin(robot_action_client_node);
-}
